// src/widgets/ArticleEditForm/model/consts/consts.ts
export enum ValidateNewArticleError {
    INCORRECT_TITLE = 'INCORRECT_TITLE',
    INCORRECT_SUBTITLE = 'INCORRECT_SUBTITLE',
    INCORRECT_IMAGE = 'INCORRECT_IMAGE',
    INCORRECT_TYPE = 'INCORRECT_TYPE',
    INCORRECT_IMAGE_URL = 'INCORRECT_IMAGE_URL',
    INCORRECT_BLOCKS = 'INCORRECT_BLOCKS',
    NO_DATA = 'NO_DATA',
    NO_USER = 'NO_USER',
    SERVER_ERROR = 'SERVER_ERROR',
}